+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 **Fixes and improvements**
 
+* Allow configuring anonymization with command-line options (in addition to the config file)
+  [#36](https://github.com/Ch00k/orthanc-cli/pull/36)
 
 ## [0.1.1](https://github.com/Ch00k/orthanc-cli/compare/0.1.0...0.1.1) 2020-12-20